remove trycatch do listener de sessão deletada e adiciona verificação de entidade no listener

+++ src/EventListener/SessionDeletedListener.php
@@ -4,6 +4,7 @@ namespace App\EventListener;
 
 use Doctrine\ORM\EntityManagerInterface;
 use EasyCorp\Bundle\EasyAdminBundle\Event\BeforeEntityDeletedEvent;
+use Symfony\Component\HttpFoundation\Session\Session;
 
 class SessionDeletedListener
 {
@@ -15,10 +16,10 @@ class SessionDeletedListener
 
     public function __invoke(BeforeEntityDeletedEvent $event): void
     {
-        try {
+        if($event->getEntityInstance() instanceof Session){
             $sid = $event->getEntityInstance()->getSession();
             $this->entityManager->getConnection()
                 ->executeStatement('DELETE FROM public.sessions WHERE sessions.sess_id = :sid', ["sid" => $sid]);
-        } catch (\Exception $e){}
+        }
     }
 }
